@@ -1,5 +1,6 @@
 {
     "name": "bukkusherufu-api",
+    "type": "module",
     "version": "1.0.0",
     "description": "Bukkusherufu API",
     "main": "src/index.js",
@@ -16,7 +17,7 @@
         "nanoid": "^3.3.6"
     },
     "devDependencies": {
-        "eslint": "^8.43.0",
+        "eslint": "^8.44.0",
         "eslint-config-standard": "^17.1.0",
         "eslint-plugin-import": "^2.27.5",
         "eslint-plugin-n": "^16.0.1",
