@@ -1,3 +1,3 @@
 const books = []
 
-module.exports = books
+export default books
