@@ -1,7 +1,7 @@
-require('dotenv').config()
-const Hapi = require('@hapi/hapi')
-const process = require('process')
-const bookRoutes = require('./routes/bookRoute')
+import 'dotenv/config'
+import Hapi from '@hapi/hapi'
+import process from 'process'
+import bookRoutes from './routes/bookRoute.js'
 
 const init = async () => {
     const server = Hapi.server({
